+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## master
+-   `PollInterceptor` refactored into `Throttler`
+-   `FixedIntervalThrottler` is becomes default `Throttler` with an interval of 10s
+
 ## v0.4.0 (05/09/2020)
 -   Timestamp offset filtering
 -   Breaking changes on SPI `HttpRequestFactory` caused by introducing `Offset` class in the model